+++ src/app/dao/tag.go
@@ -2,13 +2,13 @@ package dao
 
 // Tag tagsテーブルのDAO
 type Tag struct {
-	ID     int    `db:"id"`
-	UserID int    `db:"user_id"`
+	ID     int64  `db:"id"`
+	UserID int64  `db:"user_id"`
 	Name   string `db:"name"`
 }
 
 // TagBook tag_booksテーブルのDAO
 type TagBook struct {
-	BookID int `db:"book_id"`
-	TagID  int `db:"tag_id"`
+	BookID int64 `db:"book_id"`
+	TagID  int64 `db:"tag_id"`
 }
