+++ src/app/dao/user.go
@@ -2,7 +2,7 @@ package dao
 
 // User usersテーブルのDAO
 type User struct {
-	ID       int    `db:"id"`
+	ID       int64  `db:"id"`
 	Username string `db:"username"`
 	Password string `db:"password"`
 }
